+++ common/src/main/java/dev/ftb/packcompanion/features/CommonFeature.java
@@ -1,0 +1,17 @@
+package dev.ftb.packcompanion.features;
+
+public abstract class CommonFeature {
+    public void setup() {
+        if (isDisabled()) {
+            return;
+        }
+
+        initialize();
+    }
+
+    public abstract void initialize();
+
+    public boolean isDisabled() {
+        return true;
+    }
+}